requirements.txt
pyaml==16.12.2
gitpython==2.1.8
svn==0.3.44
docopt==0.6.2
pygithub==1.39
python-gitlab=1.3.0
pytest-cov
mock; python_version < '3'
